REF: consolidate get_indexer boilerplate
